+++ lib/src/utility/popup_manager/popup_manager.dart
@@ -19,7 +19,8 @@ final class LoaderRoute extends DialogRoute<void> {
 @immutable
 final class PopupManager {
   /// It is recommended to use root navigator's key
-  PopupManager(GlobalKey<NavigatorState> navigatorKey) : _state = navigatorKey.currentState;
+  PopupManager(GlobalKey<NavigatorState> navigatorKey)
+      : _state = navigatorKey.currentState;
 
   /// It is recommended to set rootNavigator: true eg. Navigator.of(context, rootNavigator: true)
   PopupManager.withState(this._state);
@@ -29,7 +30,11 @@ final class PopupManager {
 
   /// Shows loader dialog
   /// Provide [id] id if you have multiple loaders and want to close a specific one
-  void showLoader({String? id, bool barrierDismissible = false, WidgetBuilder? widgetBuilder}) {
+  void showLoader({
+    String? id,
+    bool barrierDismissible = false,
+    WidgetBuilder? widgetBuilder,
+  }) {
     assert(_state != null, 'Tried to show loader but navigatorState was null.');
 
     assert(
@@ -48,7 +53,7 @@ final class PopupManager {
     );
 
     _routes.add(route);
-    _state!.push(route);
+    _state.push(route);
   }
 
   /// If [id] is provided closes loader with given [id]